requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.40
chromadb>=0.4.0
pypdf2>=3.0.0
python-docx>=0.8.11
tavily-python>=0.3.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
langchain-groq>=0.1.0